+++ src/app/layouts/admin-layout/admin-layout.module.ts
@@ -1,20 +1,18 @@
-import { NgModule } from "@angular/core";
-import { HttpClientModule } from "@angular/common/http";
-import { RouterModule } from "@angular/router";
-import { CommonModule } from "@angular/common";
-import { FormsModule } from "@angular/forms";
+import { NgModule } from '@angular/core';
+import { HttpClientModule } from '@angular/common/http';import { RouterModule } from '@angular/router';
+import { CommonModule } from '@angular/common';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
-import { AdminLayoutRoutes } from "./admin-layout.routing";
-import { DashboardComponent } from "../../pages/dashboard/dashboard.component";
-import { IconsComponent } from "../../pages/icons/icons.component";
-import { MapComponent } from "../../pages/map/map.component";
-import { NotificationsComponent } from "../../pages/notifications/notifications.component";
-import { UserComponent } from "../../pages/user/user.component";
-import { TablesComponent } from "../../pages/tables/tables.component";
-import { TypographyComponent } from "../../pages/typography/typography.component";
-// import { RtlComponent } from "../../pages/rtl/rtl.component";
+import { ClipboardModule } from 'ngx-clipboard';
 
-import { NgbModule } from "@ng-bootstrap/ng-bootstrap";
+import { AdminLayoutRoutes } from './admin-layout.routing';
+import { DashboardComponent } from '../../pages/dashboard/dashboard.component';
+import { IconsComponent } from '../../pages/icons/icons.component';
+import { MapsComponent } from '../../pages/maps/maps.component';
+import { UserProfileComponent } from '../../pages/user-profile/user-profile.component';
+import { TablesComponent } from '../../pages/tables/tables.component';
+import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+// import { ToastrModule } from 'ngx-toastr';
 
 @NgModule({
   imports: [
@@ -23,16 +21,15 @@ import { NgbModule } from "@ng-bootstrap/ng-bootstrap";
     FormsModule,
     HttpClientModule,
     NgbModule,
+    ClipboardModule
   ],
   declarations: [
     DashboardComponent,
-    UserComponent,
+    UserProfileComponent,
     TablesComponent,
     IconsComponent,
-    TypographyComponent,
-    NotificationsComponent,
-    MapComponent,
-    // RtlComponent
+    MapsComponent
   ]
 })
+
 export class AdminLayoutModule {}