+++ src/app/layouts/admin-layout/admin-layout.routing.ts
@@ -1,23 +1,15 @@
-import { Routes } from "@angular/router";
+import { Routes } from '@angular/router';
 
-import { DashboardComponent } from "../../pages/dashboard/dashboard.component";
-import { IconsComponent } from "../../pages/icons/icons.component";
-import { MapComponent } from "../../pages/map/map.component";
-import { NotificationsComponent } from "../../pages/notifications/notifications.component";
-import { UserComponent } from "../../pages/user/user.component";
-import { TablesComponent } from "../../pages/tables/tables.component";
-import { TypographyComponent } from "../../pages/typography/typography.component";
-import { GpaComponent } from "src/app/pages/gpa/gpa.component";
-// import { RtlComponent } from "../../pages/rtl/rtl.component";
+import { DashboardComponent } from '../../pages/dashboard/dashboard.component';
+import { IconsComponent } from '../../pages/icons/icons.component';
+import { MapsComponent } from '../../pages/maps/maps.component';
+import { UserProfileComponent } from '../../pages/user-profile/user-profile.component';
+import { TablesComponent } from '../../pages/tables/tables.component';
 
 export const AdminLayoutRoutes: Routes = [
-  { path: "dashboard", component: DashboardComponent },
-  { path: "gpa", component: GpaComponent },
-  { path: "icons", component: IconsComponent },
-  { path: "maps", component: MapComponent },
-  { path: "notifications", component: NotificationsComponent },
-  { path: "user", component: UserComponent },
-  { path: "tables", component: TablesComponent },
-  { path: "typography", component: TypographyComponent },
-  // { path: "rtl", component: RtlComponent }
+    { path: 'dashboard',      component: DashboardComponent },
+    { path: 'user-profile',   component: UserProfileComponent },
+    { path: 'tables',         component: TablesComponent },
+    { path: 'icons',          component: IconsComponent },
+    { path: 'maps',           component: MapsComponent }
 ];