+++ src/app/layouts/auth-layout/auth-layout.module.ts
@@ -4,19 +4,19 @@ import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { AuthLayoutRoutes } from './auth-layout.routing';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { HttpClientModule } from "@angular/common/http";
 
-import { RtlComponent } from '../../pages/rtl/rtl.component';
+import { LoginComponent } from '../../pages/login/login.component';
+import { RegisterComponent } from '../../pages/register/register.component';
 @NgModule({
   imports: [
     CommonModule,
     RouterModule.forChild(AuthLayoutRoutes),
-    FormsModule,
-    HttpClientModule,
-    NgbModule
+    FormsModule
+    // NgbModule
   ],
   declarations: [
-    RtlComponent,
+    LoginComponent,
+    RegisterComponent
   ]
 })
 export class AuthLayoutModule { }