+++ src/app/layouts/auth-layout/auth-layout.routing.ts
@@ -1,8 +1,9 @@
 import { Routes } from '@angular/router';
 
-import { RtlComponent } from '../../pages/rtl/rtl.component';
-
+import { LoginComponent } from '../../pages/login/login.component';
+import { RegisterComponent } from '../../pages/register/register.component';
 
 export const AuthLayoutRoutes: Routes = [
-    { path: 'rtl',          component: RtlComponent },
+    { path: 'login',          component: LoginComponent },
+    { path: 'register',       component: RegisterComponent }
 ];