+++ src/app/pages/icons/icons.component.ts
@@ -1,11 +1,15 @@
-import { Component, OnInit } from "@angular/core";
+import { Component, OnInit } from '@angular/core';
 
 @Component({
-  selector: "app-icons",
-  templateUrl: "icons.component.html"
+  selector: 'app-icons',
+  templateUrl: './icons.component.html',
+  styleUrls: ['./icons.component.scss']
 })
 export class IconsComponent implements OnInit {
-  constructor() {}
 
-  ngOnInit() {}
+  public copy: string;
+  constructor() { }
+
+  ngOnInit() {
+  }
 }