+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.4.0] - 2022-04-1
+### Updates
+- update to Angular 13
+- update all dependencies to match Angular 13 version
+
 ## [1.3.0] - 2021-07-15
 ### Updates
 - update to Angular 12
@@ -5,15 +10,21 @@
 
 When you build and serve your app it is possible that some warnings to appear on your terminal. Those will NOT affect your product.
 
-## [1.2.0] - 2020-12-18
+## [1.2.0] - 2020-10-02
 ### Updates
-- update to Angular 11
-- update all dependencies to match Angular 11 version
+- update to Angular 10
+- update all dependencies to match Angular 10 version
 
-## [1.1.0] - 2020-03-10
+
+## [1.1.0] - 2020-03-06
 ### Updates
 - update to Angular 9
 - update all dependencies to match Angular 9 version
 
-## [1.0.0] 2019-07-19
+
+## [1.0.1] - 2019-08-05
+### Bug fixing
+- fix datepicker style
+
+## [1.0.0] - 2019-03-14
 ### Initial Release